% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVar.R
\name{getVar}
\alias{getVar}
\title{the function will get the variables selected}
\usage{
getVar(
  rdCVnetObject,
  option = c("quantile", "fitness"),
  fit_curve = "loess",
  span = 1,
  k = 5,
  outlier = "none",
  robust = 0.05,
  quantile = 0.25
)
}
\arguments{
\item{rdCVnetObject}{a object from rdCVnet}

\item{option}{quantile or fitness}

\item{fit_curve}{gam or loess}

\item{span}{for smooth curve:  how smooth the curve need to be}

\item{k}{for smooth curve:using the gam method}

\item{outlier}{if remove ourlier variables or not "none","IRQ", "residual"}

\item{robust}{If select_variables_by fitness, then it is robust}

\item{quantile}{range from 0 to 0.5. When select_variables_by quantile, this value represent the first quantile.}
}
\value{
a rdCVnet object
}
\description{
the function will get the variables selected
}
